[WIP] fix, feat (whole crate): code org and ftp mod
* reorganizing code into lib and modules
* adding ftp support
