+++ Cargo.toml
@@ -8,6 +8,7 @@ reqwest = "0.7.3"
 indicatif = "0.7.0"
 console = "0.5.0"
 clap = "2.26.0"
+ftp = "2.2.1"
 
 [dev-dependencies]
 assert_cli = "0.4"
