+++ src/lib.rs
@@ -1,0 +1,8 @@
+extern crate ftp;
+extern crate clap;
+extern crate console;
+extern crate reqwest;
+extern crate indicatif;
+
+pub mod http;
+pub mod rgetftp;
